+++ src/model_signing/_cli.py
@@ -211,6 +211,27 @@ def _sign() -> None:
         "from OIDC flow or from the environment."
     ),
 )
+@click.option(
+    "--oauth_force_oob",
+    is_flag=True,
+    default=False,
+    help=(
+        "Force an out-of-band OAuth flow and do not automatically start "
+        "the default web browser."
+    ),
+)
+@click.option(
+    "--client_id",
+    type=str,
+    metavar="ID",
+    help="The custom OpenID Connect client ID to use during OAuth2",
+)
+@click.option(
+    "--client_secret",
+    type=str,
+    metavar="SECRET",
+    help="The custom OpenID Connect client secret to use during OAuth2",
+)
 def _sign_sigstore(
     model_path: pathlib.Path,
     ignore_paths: Iterable[pathlib.Path],
@@ -218,7 +239,10 @@ def _sign_sigstore(
     signature: pathlib.Path,
     use_ambient_credentials: bool,
     use_staging: bool,
+    oauth_force_oob: bool,
     identity_token: Optional[str] = None,
+    client_id: Optional[str] = None,
+    client_secret: Optional[str] = None,
 ) -> None:
     """Sign using Sigstore (DEFAULT signing method).
 
@@ -241,6 +265,9 @@ def _sign_sigstore(
             use_ambient_credentials=use_ambient_credentials,
             use_staging=use_staging,
             identity_token=identity_token,
+            force_oob=oauth_force_oob,
+            client_id=client_id,
+            client_secret=client_secret,
         ).set_hashing_config(
             model_signing.hashing.Config().set_ignored_paths(
                 paths=list(ignore_paths) + [signature],