+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All versions prior to 1.0.0 are untracked.
 - Adjust model name when signing and verifying in model_path '.' and other fixes related to file paths
 - Fix and test the sharded file hasher
 - Added tests for verifying signatures created with v0.3.1
+- cli: `model_signing sign` now supports the `--oauth_force_oob` option (default: False)
+- Added support for specifying `--client_id` and `--client_secret` for OIDC authentication.
 
 ## [1.0.1] - 2024-04-18